requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
httpx==0.25.1
sqlmodel==0.0.14
python-dateutil==2.8.2
cartesia==1.0.9
twilio==8.10.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
